refactor: change type for Id from int to long

+++ backend/Domain/Entities/Check.cs
@@ -4,7 +4,7 @@ namespace Checklist.Domain.Entities;
 
 public class Check : AuditEntity
 {
-    public new int Id { get; set; }
+    public new long Id { get; set; }
     public string CheckName { get; set; } = string.Empty;
     public string StartTime { get; set; } = string.Empty;
     public string EndTime { get; set; } = string.Empty;
